product-automation/requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openai>=2.8.0
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0

